Rename JS class 'searchPerson' to 'PersonSearcher'
